+++ compiler+runtime/src/cpp/jank/runtime/core.cpp
@@ -2,7 +2,7 @@
 #include <jank/runtime/visit.hpp>
 #include <jank/runtime/behavior/nameable.hpp>
 #include <jank/runtime/behavior/derefable.hpp>
-#include <jank/runtime/behavior/ref.hpp>
+#include <jank/runtime/behavior/ref_like.hpp>
 #include <jank/runtime/context.hpp>
 #include <jank/runtime/sequence_range.hpp>
 #include <jank/util/fmt.hpp>
